elfloader, RISC-V: Move the elfloader during boot
The SBI implementation usually loads the elfloader where the kernel
would prefer to also be loaded. Allow the elfloader to move itself to a
higher address before loading the kernel and user images.
This requires ensuring that the stacks are not overwritten during the
merge, and so must be placed before the _archive_cpio section in the
linker script.

Signed-off-by: Kent McLeod <kent@kry10.com>